+++ lib/src/internal/main/chat_cache/channel/channel_cache.dart
@@ -1,5 +1,6 @@
 // Copyright (c) 2023 Sendbird, Inc. All rights reserved.
 
+import 'package:collection/collection.dart';
 import 'package:sendbird_chat_sdk/src/internal/main/chat_cache/cache_service.dart';
 import 'package:sendbird_chat_sdk/src/internal/main/chat_cache/channel/channel_cache_unit.dart';
 import 'package:sendbird_chat_sdk/src/internal/main/chat_cache/channel/meta_data_cache.dart';
@@ -44,6 +45,13 @@ class ChannelCache implements CacheStorage {
   @override
   void deleteAll() {
     _channelCacheMap.clear();
+  }
+
+  List<BaseChannel> getCachedChannels() {
+    return _channelCacheMap.values
+        .map((e) => e.channel)
+        .whereNotNull()
+        .toList();
   }
 
   @override